webVid_dataset/requirements.txt
opencv-python==4.6.0.66
Pillow==9.3.0
tqdm==4.64.1